Use wide-word Word128 to implement IPv6 addresses. (#52)
Use wide-word Word128 to implement IPv6 addresses. This change provides a number of new typeclass instances for free, and eliminates a bunch of code. Note that in a few cases, we still operate on the component Word64 values of the Word128 because it's significantly faster. Bump major version since this is a breaking change.