More efficient parsers and builders
This commit adds bytesmith and small-bytearray-builder to the dependency list. The upside
of this is that we can encode and decode more efficiently than we previously could. The
downside is that we can only support GHC 8.6 and up. This change include a major version
bump. The bump is not necessary, but it seems prudent to provide some extra space in the
version numbers so that additions to the API can be backported if someone is still using a
slightly older version of GHC.

This also includes several additions to the test suite and benchmark suite so that it is possible
to verify that the new encode and decode functions are correct and well performing.